Add an UI mock for storage management
Currently mquery doesn't manage its own files, which is weird given its goals.
We didn't even realise this, because we configured reindexing once, and forgot
about it. In practice though, mquery should keep track of configured folders
and automatically reindex them - without someone hacking a cron job for this.

This PR is only an UI mock, but I still want to merge this (after getting a
proper feedback of course) - for once I'll try to avoid dumping a 500 line
diff on the reviewer. It's OK to merge because it's behind a "feature toggle"
aka NODE_ENV being production - so it won't affect non-developers.

By the way, this is a second iteration. The first one was lost, because I
kept my code in /tmp and did a hasty reboot. Argh.