Fix service startup order in docker-compose ymls (#85)
* Fix service startup order in docker-compose ymls

Try to control service startup order. Otherwise they keep logging connection
errors for a first few seconds. This is not usually a problem (services
will connect eventually), but sifting through irrelevant errors when something
doesn't work may be confusing

This solution is not perfect (it doesn't check if service is ready, only
if the container started running), but it's better than nothing.

* Webs also depend on ursadb