Add filter plugins
It's currently not easily possible to remove files from mquery. As a
workaround, users can create their own blacklists for files that
should be ignored (for example, using a list of files that
should be ignored, or regular expression for file types that are
not interesting).

This also potentially allows us to query some database to check for
additional constraints on files (for example, cross check with mwdb).