Auto-restart the services on crash
Right now, when one of the services crashes (and by `one of the services` i
mean ursadb running OOM when there are too many workers) it's not started
again, and the user has to restart it manually.

Of course services crashing is a pretty bad thing (and should not happen),
but having a database lose a single task and restart is still better than
user having to restart it manually.