Add a configurable about page (#341)
This PR adds a customisable "about" page. It's not very customisable, because all the information will have to fit into a config file, but that should be enough to give users a few URLs or explain what's going on.

This subpage is optional (will be optional in the final PR).

I've decided to allow HTML (it's trusted - it comes from the server) instead of just plain text.